split cowrie/ssh/session, move shell to shell/
